Moved ensure_gio_init() into a common library
GioUtils has been created. It now has a function that both DarkModeGnome.cpp and GnomeThemeUtils.cpp can access to make sure that Gio::init() is called only once.
